+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 - all: update logo
 - all: move main into project root
 - all: detach console when running on Windows
+- agent: update to use named pipe on Windows
 - otp: ensure decoded key is padded
 - otp: fix padding issue for the 2fa code
+
+- deps add:
+    - gopkg.in/natefinch/npipe.v2 v2.0.0-20160621034901-c1b8fa8bdcce
 
 ## 0.21.2 - 28 January 2024